+++ src/main/java/group7/springmvc/repository/RoleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import group7.springmvc.model.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long>{
+public interface RoleRepository extends JpaRepository<Role, Byte>{
 	Role findByValue(String value);
 }
